requirements.txt
ultralytics
opencv-python
numpy
streamlit
pillow
